requirements.txt
scipy>=1.4.0
numpy>=1.19.2
Keras==2.10.0
matplotlib>=3.1.2
opencv_python>=4.1.2.30
tensorflow_gpu==2.10.0
tqdm>=4.60.0
Pillow>=8.2.0
h5py>=3.1.0

# Keras==2.10.0
# tensorflow==2.10.0